+++ graphquery/core/models/lpg_mmap/index_file.hpp
@@ -44,7 +44,7 @@ namespace graphquery::database::storage
          ***************************************************************/
         struct SIndexEntry_t
         {
-            uint32_t id     = {};
+            uint64_t id     = {};
             uint32_t offset = END_INDEX;
         };
 
@@ -58,9 +58,9 @@ namespace graphquery::database::storage
         CDiskDriver & get_file() noexcept;
         inline void store_metadata() noexcept;
         inline SRef_t<SIndexMetadata_t> read_metadata() noexcept;
-        inline SRef_t<SIndexEntry_t> read_entry(uint32_t offset) noexcept;
+        inline SRef_t<SIndexEntry_t> read_entry(int64_t offset) noexcept;
         void open(std::filesystem::path path, std::string_view file_name, bool create) noexcept;
-        void store_entry(uint32_t id, uint32_t offset) noexcept;
+        void store_entry(uint64_t id, int64_t offset) noexcept;
 
       private:
         CDiskDriver m_file;
@@ -86,7 +86,7 @@ graphquery::database::storage::CIndexFile::store_metadata() noexcept
 }
 
 inline graphquery::database::storage::SRef_t<graphquery::database::storage::CIndexFile::SIndexEntry_t>
-graphquery::database::storage::CIndexFile::read_entry(const uint32_t offset) noexcept
+graphquery::database::storage::CIndexFile::read_entry(const int64_t offset) noexcept
 {
     static const auto base_addr  = utils::atomic_load(&read_metadata()->index_list_start_addr);
     static const auto index_size = utils::atomic_load(&read_metadata()->index_size);
@@ -95,9 +95,9 @@ graphquery::database::storage::CIndexFile::read_entry(const uint32_t offset) noe
 }
 
 inline void
-graphquery::database::storage::CIndexFile::store_entry(const uint32_t id, const uint32_t offset) noexcept
+graphquery::database::storage::CIndexFile::store_entry(const uint64_t id, const int64_t offset) noexcept
 {
-    auto index_ptr     = read_entry(id);
+    auto index_ptr     = read_entry(static_cast<int64_t>(id));
     const auto index_c = utils::atomic_load(&read_metadata()->index_c);
 
     utils::atomic_store(&index_ptr->id, id);
